+++ src/value.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Clone)]
 pub enum Value {
     Str(String),
-    Int(i32),
+    Int(usize),
     Bool(bool),
     None
 }
